requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
yt-dlp>=2025.08.11	
openai-whisper
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.0,<3.0.0
ffmpeg-python==0.2.0
psutil==5.9.6
groq== 0.30.0
python-dotenv==1.0.0
firebase-admin==6.2.0
boto3==1.34.0
PyMuPDF>=1.23.0
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
httpx>=0.25.0
user-agents>=2.2.0


dropbox>=11.36.0
soundfile>=0.12.1
pyttsx3>=2.90
gtts>=2.3.0
watchdog>=3.0.0
email-validator>=2.1.0
spacy>=3.7.0,<3.8.0
thinc>=8.3.4,<8.4.0
blis>=0.7.0,<1.0.0